Bug 6976 Close loophole allowing borrowers extra holds
The check that the max number of reserves has
not been reached needs to take account of the fact
that we may have added enouch reserves to
reach the limit while this script is running.
Add a check against the incrementing count

The variable canreserve was only used when looping
through multiple holds but was not being
reinitialized on every iteration
make variable local to loop it is used in and
initialize to zero

Some numeric comparisons were using a string operator
fix them before strange bugs manifest themselves

The loophole manifests thus:
borrowers are only allowed Y holds.  If holds are done individually, it is
obeying that law i.e. X has held Y separate books – when I tried to do a 6th
book it told me that I could not place a hold.  However, if a borrower checks
the catalogue, gets the intial list up and clicks items on the “results” box
and then clicks “place hold” they can hold more than Y items